+++ augur/io/__init__.py
@@ -4,5 +4,5 @@
 # To use functions internally, import directly from the submodule.
 from .file import open_file  # noqa: F401
 from .metadata import read_metadata  # noqa: F401
-from .sequences import read_sequences, write_sequences  # noqa: F401
+from .sequences import read_sequences, write_sequences, load_features  # noqa: F401
 from .strains import read_strains  # noqa: F401