+++ augur/utils.py
@@ -7,6 +7,7 @@ import pandas as pd
 from collections import OrderedDict
 from io import RawIOBase
 from shlex import quote as shquote
+from typing import List, Union
 from .__version__ import __version__
 
 from augur.data import as_file
@@ -20,16 +21,33 @@ from augur.util_support.color_parser import ColorParser
 from augur.util_support.node_data_reader import NodeDataReader
 
 
-def augur():
+def augur(shell: bool = True) -> Union[str, List[str]]:
     """
     Locate how to re-invoke ourselves (_this_ specific Augur).
+
+    This function returns the appropriate command to re-invoke the current Augur 
+    installation, useful for subprocess calls that need to launch Augur commands.
+    The format of the returned command depends on the shell parameter, which 
+    mirrors the identically named parameter to subprocess.run.
+
+    Parameters
+    ----------
+    shell
+        Controls the format of the returned command. When True, returns a string
+        suitable for shell=True in subprocess.run. When False, returns a list of
+        arguments suitable for shell=False in subprocess.run. Default is True.
     """
     if sys.executable:
-        return f"{shquote(sys.executable)} -m augur"
+        args = [shquote(sys.executable), "-m", "augur"]
     else:
         # A bit unusual we don't know our own Python executable, but assume we
         # can access ourselves as the ``augur`` command.
-        return f"augur"
+        args = ["augur"]
+
+    if shell:
+        return " ".join(args)
+    else:
+        return args
 
 
 def get_json_name(args, default=None):
@@ -179,7 +197,7 @@ class AugurJSONEncoder(json.JSONEncoder):
 
 
 def load_features(*args, **kwargs):
-    print_err("WARNING: augur.utils.load_features has been moved to augur.io.sequences.load_features.")
+    print_err("DEPRECATION WARNING: augur.utils.load_features has been moved to augur.io.load_features. The old alias will be removed in a future major version.")
     from augur.io.sequences import load_features as new_location
     return new_location(*args, **kwargs)